rename: 컴포넌트 ImageUpload 의 위치를 ui 디렉토리에서 product-register 로 이동 및 @web 별칭으로 import 문 적용 (#164)
