+++ apps/web/components/product-register/ProductRegisterForm.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from '@repo/ui/components';
 
-import ImageUpload from '@/components/ui/ImageUpload';
+import { ImageUpload } from '@/components/product-register';
 
 import { useProductForm } from '@/hooks/products';
