+++ apps/web/components/product-register/index.ts
@@ -5,3 +5,4 @@ export { default as MinimumPriceField } from './MinimumPriceField';
 export { default as DecreaseUnitField } from './DecreaseUnitField';
 export { default as LocationInfoField } from './LocationInfoField';
 export { default as ProductRegisterForm } from './ProductRegisterForm';
+export { default as ImageUpload } from './ImageUpload';
